requirements.txt
PyGame
Numba
PyGLM
toml
loguru
kivy